==> r-code/01_2_iNext.3d_butterfly.Rmd <==
---
title: "Comparisons of sample coverages using iNext"
author: "Bernd Panassiti"
date: 'created: 07.06.2023, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output:
  pdf_document: default
  html_document: default
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=FALSE,                       # shows r-code
                results='hide',                    # don't display results
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                fig.show ='hide',
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```


For installation of iNEXT.beta3D:

https://github.com/AnneChao/iNEXT.beta3D


```{r load,, warning=FALSE}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load("tidyverse","vegan","reshape2","magrittr","caret","ggplot2",
       "iNEXT","iNEXT.beta3D","data.table",
       "FD","gawdis")
source("r-code/00_functions.R")
load("data/rawdata.RData")
```

# Introduction
The co-occurrence dataset contains data on **pol**linators (wildbutterflys, butterflies, hoverflies) and **phy**tophagous species (planthoppers).  
  
The following data are available:  

* year: **pol** (2021,2022), **phy** (2021)
* survey periods: **pol** (1,2), **phy** (2)
* treatments: control, early, late

Create list to save the following datasets:
* alpha TD
* alpha PD
* alpha FD
* beta TD
* beta PD
* beta FD


# Warning message due to "future" package
# https://github.com/satijalab/seurat/issues/3622
options(future.rng.onMisuse="ignore")



# Data preparation
# TD dataframe
```{r}
dfButterfly = butterfly.abund %>%
  dplyr::group_by(species,plot_id)  %>%
  dplyr::summarise(count=sum(count))

butterflyAbuTD = data.frame(reshape::cast(dfButterfly,species~plot_id,fill=0))
rownames(butterflyAbuTD) = butterflyAbuTD[,1]
butterflyAbuTD = butterflyAbuTD[,-1]
```


# PD
## Check for same species names with trait matrix
```{r}
butterflyAbuPD = butterflyAbuTD

iTree = butterflyTreeComm
iTree$node.label =  paste("match",rep(1:length(iTree$node.label)),sep="_")


# which species in in butterfly dataframe are not in traits
butterflyAbuPDTree = butterflyAbuPD[-which(!rownames(butterflyAbuPD) %in% iTree$tip.label),]

# cbind(1:ncol(butterflyAbuPD),apply(butterflyAbuPD,2,function(x) length(which(x>0))))
# remove plots 0 or 1 species

butterflyAbuPD = butterflyAbuPDTree[,-which(apply(butterflyAbuPDTree,2,function(x) length(which(x>0)))<2)]

# remove plots 1 species
plots0speciesPD = names(which(apply(butterflyAbuPDTree,2,function(x) length(which(x>0)))==0))
plots1speciesPD = names(which(apply(butterflyAbuPDTree,2,function(x) length(which(x>0)))==1))
```


# FD
## Check for same species names with trait matrix
```{r}
butterflyAbuFD = butterflyAbuTD

iTraits = butterflyTraits
iTraits$species = gsub(" ","_",butterflyTraits$species) # recode species names with underline

# which species in in butterfly dataframe are not in traits
butterflyAbuFDTrait = butterflyAbuFD[-which(!rownames(butterflyAbuFD) %in% iTraits$species),]

# cbind(1:ncol(butterflyAbuFD),apply(butterflyAbuFD,2,function(x) length(which(x>0))))
# remove plots 1 species

butterflyAbuFD = butterflyAbuFDTrait[,-which(apply(butterflyAbuFDTrait,2,function(x) length(which(x>0)))<2)]

plots1speciesFD = names(which(apply(butterflyAbuFDTrait,2,function(x) length(which(x>0)))==1))
plots0speciesFD = names(which(apply(butterflyAbuFDTrait,2,function(x) length(which(x>0)))==0))


# plots0speciesFD
# "A002_1_2022_early_intensive"


# which species in traits are not in butterfly dataframe
iTraits = iTraits[-which(!iTraits$species %in% rownames(butterflyAbuFD)),]


# rownames(iTraits) = iTraits$species
# butterflyAbu_distM <- FD::gowdis(iTraits[,-c(1:2)])
butterflyAbu_distM <-data.frame(StatMatch::gower.dist(iTraits[,-c(1:2)]))
colnames(butterflyAbu_distM) =  iTraits$species
rownames(butterflyAbu_distM) =  iTraits$species
```


# Save all datasets in list
```{r}
butterflyInput = list(
  butterflyAbuTD = butterflyAbuTD,
  butterflyAbuPD = butterflyAbuPD, # not columns with less than 2 species and only species that are also in tree
  butterflyAbuFD = butterflyAbuFD, # not columns with less than 2 species and only species that are also in trait matrix
  butterflyAbuPDTree = butterflyAbuPDTree, # all columns but only species that are also in tree
  butterflyAbuFDTrait = butterflyAbuFDTrait, # all columns but only species that are also in trait matrix
  iTree = iTree,
  iDistM = butterflyAbu_distM,
  plots1speciesPD = plots1speciesPD,
  plots0speciesPD = plots0speciesPD,
  plots1speciesFD = plots1speciesFD,
  plots0speciesFD = plots0speciesFD)
```

# Calculate estimates
##TD
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2023)

out.TD <- iNEXT.3D::estimate3D(data = butterflyInput$butterflyAbuTD, 
               diversity = "TD", 
               q = 0, 
               datatype = "abundance", 
               base = "coverage", level = 0.7,
               nboot = 30)
```


## PD
out.PD <- iNEXT.3D::estimate3D(data = butterflyInput$butterflyAbuPD, diversity = "PD", 
               q = 0, datatype = "abundance", 
               PDtree = butterflyInput$iTree,
               base = "coverage", level = 0.7,
               nboot = 30)


```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2023)


outPDtmp = list()
for (i in 1:ncol(butterflyInput$butterflyAbuPD)){
dat = data.frame(butterflyInput$butterflyAbuPD[,i])
rownames(dat) = rownames(butterflyInput$butterflyAbuPD)

try(outPDtmp[[i]] <- iNEXT.3D::estimate3D(data = dat, diversity = "PD", 
               q = 0, datatype = "abundance", 
               PDtree = butterflyInput$iTree,
               base = "coverage", level = 0.7,
               nboot = 30))


try(
  outPDtmp[[i]]$Assemblage <- colnames(butterflyInput$butterflyAbuPD)[i]
  ,silent=TRUE)
  
print(i)}


# Fehler in if (refC > cvrg) { : Fehlender Wert, wo TRUE/FALSE nötig ist


# generate estimates for sites with errors with set.seed(2023)
newID = which(!lengths(outPDtmp));newID
# newID
# 9 11 17 20 25 26 27 28 48 50 51 53 56
df = butterflyInput$butterflyAbuPD




outNewID=list()
for (i in 1:length(newID)){
j=0
out.id=list()

while(length(which(lengths(out.id)!=0))<1){
j=j+1

set.seed(j)
dat = data.frame(df[,newID[i]])
rownames(dat) = rownames(df)

try(out.id[[j]] <- iNEXT.3D::estimate3D(data = dat, diversity = "PD", 
               q = 0, datatype = "abundance", 
               PDtree = butterflyInput$iTree,
               base = "coverage", level = 0.7,
               nboot = 30))

}

out.id[[length(out.id)]]$Assemblage = colnames(df)[newID[i]]
outNewID[[i]] = out.id
print(newID[i])}

# combine both list

outList1 = outPDtmp[!unlist(lapply(outPDtmp,is.null))]
outList1DF = do.call(rbind, outList1)

outList2 = lapply(outNewID, function(x) x[!unlist(lapply(x,is.null))])
outList2DF = do.call(rbind, lapply(outList2, `[[`, 1))


out.PD = bind_rows(outList1DF,outList2DF)

#save seed
seedPD = data.frame(
  plot_id =colnames(butterflyInput$butterflyAbuPD)[newID],
  newSeed=unlist(lapply(outNewID, function(x) which(!unlist(lapply(x,is.null)))))
)
```




## FD
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2023)

out.FD <- iNEXT.3D::estimate3D(data = butterflyInput$butterflyAbuFD, diversity = "FD", 
               q = 0, datatype = "abundance", 
               FDdistM = butterflyInput$iDistM,
               base = "coverage", level = 0.7,
               nboot = 30)
```

```{r}
butterfly.TD = out.TD
butterfly.PD = out.PD
butterfly.FD = out.FD
```


# Combine all datasets
```{r}
out1 = out.TD %>% dplyr::mutate(diversityType ="TD")  %>%
                  dplyr::select(plot_id = Assemblage,diversityType,
                         qHill = Order.q, obs = m, 
                         estimate = qTD, sd= s.e., lowCI = qTD.LCL, highCI= qTD.UCL,
                         sampleCoverage = SC,method = Method)
                  

out2 = out.PD %>% dplyr::mutate(diversityType ="PD")  %>%
                  dplyr::select(plot_id = Assemblage,diversityType,
                         qHill = Order.q, obs = m, 
                         estimate = qPD, sd = s.e., lowCI = qPD.LCL, highCI = qPD.UCL,
                         sampleCoverage = SC,method = Method)

out3 = out.FD %>% dplyr::mutate(diversityType ="FD")  %>%
                  dplyr::select(plot_id = Assemblage,diversityType,
                         qHill = Order.q, obs = m, 
                         estimate = qFD, sd = s.e., lowCI = qFD.LCL, highCI = qFD.UCL,
                         sampleCoverage = SC,method = Method)

out = as_tibble(bind_rows(out1,out2,out3))


##FD

# plots1speciesFD 
# vegan::diversity(butterflyInput$butterflyAbuFDTrait[,plots1speciesFD[11]],"shannon")
# 0

a = length(plots1speciesFD)
out.FD1 = data.frame(plot_id=plots1speciesFD,
                     diversityType = rep("FD",a),
                     qHill = rep(0,a),
                     obs=rep(1,a),
                     estimate=rep(1,a),
                     sd=rep(NA,a),lowCI=rep(NA,a),highCI=rep(NA,a),
                     sampleCoverage = rep(0.7,a),method=rep("Observed",a))

# plots0speciesFD
a = length(plots0speciesFD)

out.FD0 = data.frame(plot_id=plots0speciesFD,diversityType = rep("FD",a),
                     qHill = rep(0,a),
                     obs=rep(0,a),
                     estimate=rep(0,a),
                     sd=rep(NA,a),lowCI=rep(NA,a),highCI=rep(NA,a),
                     sampleCoverage = rep(0.7,a),method=rep("Observed",a))


##PD

# plots1speciesPD 
a = length(plots1speciesPD)

out.PD1 = data.frame(plot_id=plots1speciesPD,diversityType = rep("PD",a),
                     qHill = rep(0,a),
                     obs=rep(1,a),
                     estimate=rep(1,a),
                     sd=rep(NA,a),lowCI=rep(NA,a),highCI=rep(NA,a),
                     sampleCoverage = rep(0.7,a),method=rep("Observed",a))

# plots0speciesPD
a = length(plots0speciesPD)

out.PD0 = data.frame(plot_id=plots0speciesPD,diversityType = rep("PD",a),
                     qHill = rep(0,a),
                     obs=rep(0,a),
                     estimate=rep(0,a),
                     sd=rep(NA,a),lowCI=rep(NA,a),highCI=rep(NA,a),
                     sampleCoverage = rep(0.7,a),method=rep("Observed",a))

#butterflyDiversity = list()
butterflyDiversity = dplyr::bind_rows(out,out.PD1,out.PD0,out.FD1,out.FD0)

butterflyDiversity$site_id = as.character(sapply(butterflyDiversity$plot_id, function(x) str_split(x,"_")[[1]][1]))
butterflyDiversity$survey_period = as.character(sapply(butterflyDiversity$plot_id, function(x) str_split(x,"_")[[1]][2]))
butterflyDiversity$year = as.integer(sapply(butterflyDiversity$plot_id, function(x) str_split(x,"_")[[1]][3]))
butterflyDiversity$grazingTime = as.character(sapply(butterflyDiversity$plot_id, function(x) str_split(x,"_")[[1]][4]))
```



# Save results
```{r}
save(butterflyInput,butterflyDiversity,
     butterfly.TD,butterfly.FD,butterfly.PD,
     seedPD,
     file="results/iNEXT3d_butterfly.RData")
```
