==> r-code/03_Q1_3_statistical_analyses_percent_change.Rmd <==
---
title: "Calculation of percent change"
author: "Bernd Panassiti"
date: 'created: 16.09.2023, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output:
  pdf_document: default
  html_document: default
editor_options: 
  chunk_output_type: console
---



```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```




# Introduction

This script provides the R-code to carry out the calculation of percent changes based on the statistical analyses to evaluate the influence of grazing time on insect diversity in alpine pastures.



# Loading of packages and data
```{r loading_data, include=FALSE}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load(tidyverse,magrittr,caret,ggplot2,DHARMa,broom.mixed,
       posterior,ggdist,tidybayes,pryr,bayesplot,ggpubr,grid,gridExtra,
       emmeans,broom, broom.mixed,
       posterior,brms,bayesplot,ggpubr,
       hrbrthemes,patchwork,# additional themes
       GGally) # ggpairs

source("r-code/00_functions.R")

load("results/modelResults.RData")
```

## Insect model - Management predictors

#### Percent change
##### Management plot bees
```{r}
# Output the conditional effects.
fit.TD = modelResultsTDManagement[[1]]
fit.PD = modelResultsPDManagement[[1]]
fit.FD = modelResultsFDManagement[[1]]

cond_mat.TD <- conditional_effects(fit.TD,"grazingTime:surveyPeriod")[[1]]
cond_mat.TD$type = rep("TD",nrow(cond_mat.TD))
cond_mat.TD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.TD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

cond_mat.PD <- conditional_effects(fit.PD,"grazingTime:surveyPeriod")[[1]] 
cond_mat.PD$type = rep("PD",nrow(cond_mat.PD))
cond_mat.PD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.PD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

cond_mat.FD <- conditional_effects(fit.FD,"grazingTime:surveyPeriod")[[1]] 
cond_mat.FD$type = rep("FD",nrow(cond_mat.FD))
cond_mat.FD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.FD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

m = bind_rows(cond_mat.TD,cond_mat.PD,cond_mat.FD)

# 3 diversity types (TD, PD, FD)
# 2 survey periods
m$value = NA
m$se_low = NA
m$se_high = NA
m[which(m$grazingTime == "control"),"value"] = 0


for (i in c("TD","PD","FD")){
for (j in c(1,2)){
  
estimateControl   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"estimate"]  
estimateEarly   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"estimate"]
estimateLate    = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"estimate"]
seControl       = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"se"]
seEarly         = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"se"]
seLate          = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"se"]

m[which(m$type==i & m$surveyPeriod == j & m$grazingTime == "control"),"se_low"]  =
  (((estimateControl-seControl)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & m$grazingTime == "control"),"se_high"] = 
  (((estimateControl+seControl)/estimateControl)-1)*100

# estimate
m[which(m$type==i & m$surveyPeriod == j & m$grazingTime == "early"),"value"]   =
  ((estimateEarly/estimateControl)-1)*100
# se low
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_low"]  = (((estimateEarly-seEarly)/estimateControl)-1)*100
# se high
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_high"] = (((estimateEarly+seEarly)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"value"]    = ((estimateLate/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_low"]   = (((estimateLate-seLate)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_high"]  = (((estimateLate+seLate)/estimateControl)-1)*100
}
}

m$surveyPeriod = plyr::revalue(m$surveyPeriod, c("1"="Summer onset", "2"="Summer peak"))
m$type = as.factor(m$type)
m$type <- factor(m$type, levels = c("TD", "PD", "FD"))
levels(m$type) <- c("Taxonomic", "Phylogenetic","Functional")

percentChangeBeeSRManagement = m
```


##### Management plot butterfly
```{r}
# Output the conditional effects
fit.TD = modelResultsTDManagement[[2]]
fit.PD = modelResultsPDManagement[[2]]
fit.FD = modelResultsFDManagement[[2]]

cond_mat.TD <- conditional_effects(fit.TD,"grazingTime:surveyPeriod")[[1]]
cond_mat.TD$type = rep("TD",nrow(cond_mat.TD))
cond_mat.TD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.TD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

cond_mat.PD <- conditional_effects(fit.PD,"grazingTime:surveyPeriod")[[1]] 
cond_mat.PD$type = rep("PD",nrow(cond_mat.PD))
cond_mat.PD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.PD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

cond_mat.FD <- conditional_effects(fit.FD,"grazingTime:surveyPeriod")[[1]] 
cond_mat.FD$type = rep("FD",nrow(cond_mat.FD))
cond_mat.FD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.FD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

m = bind_rows(cond_mat.TD,cond_mat.PD,cond_mat.FD)

# 3 diversity types (TD, PD, FD)
# 2 survey periods
m$value = NA
m$se_low = NA
m$se_high = NA
m[which(m$grazingTime == "control"),"value"] = 0



for (i in c("TD","PD","FD")){
for (j in c(1,2)){
estimateControl   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"estimate"]  
estimateEarly   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"estimate"]
estimateLate    = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"estimate"]
seControl       = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"se"]
seEarly         = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"se"]
seLate          = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"se"]

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "control"),"se_low"]  = (((estimateControl-seControl)/estimateControl)-1)*100
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "control"),"se_high"] = (((estimateControl+seControl)/estimateControl)-1)*100

# estimate
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"value"]   = ((estimateEarly/estimateControl)-1)*100
# se low
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_low"]  = (((estimateEarly-seEarly)/estimateControl)-1)*100
# se high
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_high"] = (((estimateEarly+seEarly)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"value"]    = ((estimateLate/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_low"]   = (((estimateLate-seLate)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_high"]  = (((estimateLate+seLate)/estimateControl)-1)*100
}
}

m$surveyPeriod = plyr::revalue(m$surveyPeriod, c("1"="Summer onset", "2"="Summer peak"))
m$type = as.factor(m$type)
m$type <- factor(m$type, levels = c("TD", "PD", "FD"))
levels(m$type) <- c("Taxonomic", "Phylogenetic","Functional")

percentChangeButterflySRManagement = m
```


##### Management plot leafhopper
```{r}
# Output the conditional effects.
fit.TD = modelResultsTDManagement[[3]]
fit.PD = modelResultsPDManagement[[3]]
fit.FD = modelResultsFDManagement[[3]]

cond_mat.TD <- conditional_effects(fit.TD,"grazingTime:surveyPeriod")[[1]]
cond_mat.TD$type = rep("TD",nrow(cond_mat.TD))
cond_mat.TD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.TD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

cond_mat.PD <- conditional_effects(fit.PD,"grazingTime:surveyPeriod")[[1]] 
cond_mat.PD$type = rep("PD",nrow(cond_mat.PD))
cond_mat.PD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.PD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

cond_mat.FD <- conditional_effects(fit.FD,"grazingTime:surveyPeriod")[[1]] 
cond_mat.FD$type = rep("FD",nrow(cond_mat.FD))
cond_mat.FD %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.FD %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

m = bind_rows(cond_mat.TD,cond_mat.PD,cond_mat.FD)

# 3 diversity types (TD, PD, FD)
# 2 survey periods
m$value = NA
m$se_low = NA
m$se_high = NA
m[which(m$grazingTime == "control"),"value"] = 0


for (i in c("TD","PD","FD")){
for (j in c(1,2)){
estimateControl   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"estimate"]  
estimateEarly   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"estimate"]
estimateLate    = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"estimate"]
seControl       = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"se"]
seEarly         = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"se"]
seLate          = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"se"]

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "control"),"se_low"]  = (((estimateControl-seControl)/estimateControl)-1)*100
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "control"),"se_high"] = (((estimateControl+seControl)/estimateControl)-1)*100

# estimate
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"value"]   = ((estimateEarly/estimateControl)-1)*100
# se low
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_low"]  = (((estimateEarly-seEarly)/estimateControl)-1)*100
# se high
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_high"] = (((estimateEarly+seEarly)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"value"]    = ((estimateLate/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_low"]   = (((estimateLate-seLate)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_high"]  = (((estimateLate+seLate)/estimateControl)-1)*100

}
}

m$surveyPeriod = plyr::revalue(m$surveyPeriod, c("1"="Summer onset", "2"="Summer peak"))
m$type = as.factor(m$type)
m$type <- factor(m$type, levels = c("TD", "PD", "FD"))
levels(m$type) <- c("Taxonomic", "Phylogenetic","Functional")


percentChangeLeafhopperSRManagement = m
```

##### Prepare for Excel export and number of paper
```{r}
percentChangeBeeSRManagement$group ="Bees"
percentChangeButterflySRManagement$group = "Butterflies"
percentChangeLeafhopperSRManagement$group = "Leafhoppers"

dat = bind_rows(percentChangeBeeSRManagement,
                percentChangeButterflySRManagement,
                percentChangeLeafhopperSRManagement)
dat$surveyPeriod = as.character(dat$surveyPeriod)
dat[dat$surveyPeriod=="Survey period 1","surveyPeriod"] =1
dat[dat$surveyPeriod=="Survey period 2","surveyPeriod"] =2
resultsExcelPercentChangeSRManagement = dat

# how much less in survey period 1 then 2 for treatment early
##bee
dat %>% dplyr::filter(grazingTime!="control") %>%
  dplyr::group_by(group,surveyPeriod,grazingTime)%>%
  dplyr::summarize(averageChange = mean(value))

```

   group        surveyPeriod grazingTime averageChange
   <chr>        <chr>        <fct>               <dbl>
 1 Bees         1            early              -41.0 
 2 Bees         1            late               -12.3 
 3 Bees         2            early                6.06
 4 Bees         2            late               -21.5 
 5 Butterflies  1            early               -6.16
 6 Butterflies  1            late                21.9 
 7 Butterflies  2            early               39.4 
 8 Butterflies  2            late                41.6 
 9 Planthoppers 1            early              -18.0 
10 Planthoppers 1            late                -3.32
11 Planthoppers 2            early               -6.21
12 Planthoppers 2            late                -2.51

## Vegetation model - Management predictors

#### Percent change
```{r}
# SR
fit.vegSR = modelResultsVegManagement[[1]]
cond_mat.vegSR <- conditional_effects(fit.vegSR,"grazingTime:surveyPeriod")[[1]]
cond_mat.vegSR$type = rep("vegSR",nrow(cond_mat.vegSR))
cond_mat.vegSR %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.vegSR %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

# flower
fit.vegFlower = modelResultsVegManagement[[2]]
cond_mat.vegFlower <- conditional_effects(fit.vegFlower,"grazingTime:surveyPeriod")[[1]]
cond_mat.vegFlower$type = rep("vegFlower",nrow(cond_mat.vegFlower))
cond_mat.vegFlower %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.vegFlower %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

# height
fit.vegHeight = modelResultsVegManagement[[3]]
cond_mat.vegHeight <- conditional_effects(fit.vegHeight,"grazingTime:surveyPeriod")[[1]]
cond_mat.vegHeight$type = rep("vegHeight",nrow(cond_mat.vegHeight))
cond_mat.vegHeight %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.vegHeight %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)

# biomass
fit.vegBiomass = modelResultsVegManagement[[4]]
cond_mat.vegBiomass <- conditional_effects(fit.vegBiomass,"grazingTime:surveyPeriod")[[1]]
cond_mat.vegBiomass$type = rep("vegBiomass",nrow(cond_mat.vegBiomass))
cond_mat.vegBiomass %<>% dplyr::rename(estimate = "estimate__",se="se__",lowCI = "lower__", highCI = "upper__")
cond_mat.vegBiomass %<>% dplyr::select(type,grazingTime,surveyPeriod,estimate,se,lowCI,highCI)


m = bind_rows(cond_mat.vegSR,cond_mat.vegFlower,cond_mat.vegHeight,cond_mat.vegBiomass)

# 4 plant responses (vegSR,vegFlower,vegHeight,vegBiomass)
# 2 survey periods
m$value = NA
m$se_low = NA
m$se_high = NA
m[which(m$grazingTime == "control"),"value"] = 0


for (i in c("vegSR","vegFlower","vegHeight","vegBiomass")){
for (j in c(1,2)){
  
estimateControl   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"estimate"]  
estimateEarly   = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"estimate"]
estimateLate    = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"estimate"]
seControl       = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "control"),"se"]
seEarly         = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "early"),"se"]
seLate          = m[which(m$type==i & m$surveyPeriod == j & 
                            m$grazingTime == "late"),"se"]

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "control"),"se_low"]  = (((estimateControl-seControl)/estimateControl)-1)*100
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "control"),"se_high"] = (((estimateControl+seControl)/estimateControl)-1)*100

# estimate
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"value"]   = ((estimateEarly/estimateControl)-1)*100
# se low
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_low"]  = (((estimateEarly-seEarly)/estimateControl)-1)*100
# se high
m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "early"),"se_high"] = (((estimateEarly+seEarly)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"value"]    = ((estimateLate/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_low"]   = (((estimateLate-seLate)/estimateControl)-1)*100

m[which(m$type==i & m$surveyPeriod == j & 
          m$grazingTime == "late"),"se_high"]  = (((estimateLate+seLate)/estimateControl)-1)*100
}
}



# plot
levels(m$surveyPeriod) <- c("Survey period 1", "Survey period 2")
m$type = as.factor(m$type)
m$type <- factor(m$type, levels = c("vegSR","vegFlower","vegHeight","vegBiomass"))
m$label = m$type
levels(m$label) <- c("Species richness","Number of Infloresences","Vegetation height","Biomass")

#dfPlot = m %>% dplyr::filter(type %in% c("vegSR","vegFlower","vegHeight"))
dfPlot =m

vegPlotManagement = ggplot(dfPlot, aes(x = grazingTime,y = value,fill = label))+
  xlab("Grazing times") + ylab("Percent change (%)") +
  facet_grid(.~surveyPeriod) +
  geom_hline(yintercept=0, lty=2, lwd=1, colour="black") +
  geom_errorbar(aes(ymin=se_low,
                    ymax=se_high,
                    colour=label),
                lwd=1,width=0,
                position = position_dodge(width=0.2))+
  geom_point(size=3, aes(colour=label),
             position = position_dodge(width=0.2)) +
   guides(colour=guide_legend(title=""),fill="none")+
  ggtitle("Vegetation") +
  theme(axis.title = element_text(size=18),  
        axis.line = element_line(colour="black"),
        axis.ticks = element_line(),
        axis.text = element_text(size = 18, colour="black"),
        panel.border = element_blank(),
        title = element_text(size=22),
        legend.text=element_text(size=18),
        strip.text = element_text(size=18),
        legend.position="bottom");vegPlotManagement

percentChangeVeg = m
```



## Insect model - Vegetation predictors 

#### Percent change
##### Vegetation plot bees
```{r}
fit.TD = modelResultsTDVegetation[[1]]
fit.PD = modelResultsPDVegetation[[1]]
fit.FD = modelResultsFDVegetation[[1]]

fixef.TD <- data.frame(fixef(fit.TD))
fixef.TD$variable <- rownames(fixef.TD)
fixef.TD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","year2022","elevationScaled"))
fixef.TD$type = rep("TD",nrow(fixef.TD))
fixef.TD %<>% dplyr::rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")


fixef.PD <- data.frame(fixef(fit.PD))
fixef.PD$variable <- rownames(fixef.PD)
fixef.PD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","year2022","elevationScaled"))
fixef.PD$type = rep("PD",nrow(fixef.PD))
fixef.PD %<>% dplyr::rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")

fixef.FD <- data.frame(fixef(fit.FD))
fixef.FD$variable <- rownames(fixef.FD)
fixef.FD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","year2022","elevationScaled"))
fixef.FD$type = rep("FD",nrow(fixef.FD))
fixef.FD %<>% dplyr::rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")



m = bind_rows(fixef.TD,fixef.PD,fixef.FD)
rownames(m)= 1:nrow(m)
# 3 diversity types (TD, PD, FD)
# 2 survey periods
m$value   = NA
m$se_low  = NA
m$se_high = NA

m$value= (exp(m$estimate)-1)*100
se=(exp(m$se)-1)*100
m$se_low  = m$value-se
m$se_high = m$value+se

m$surveyPeriod = 1
m[which(grepl("surveyPeriod2",m$variable)),"surveyPeriod"] = 2

# plot
m$surveyPeriod = as.factor(m$surveyPeriod)
m$surveyPeriod = plyr::revalue(m$surveyPeriod, c("1"="Summer onset", "2"="Summer peak"))
m$type = as.factor(m$type)
m$type <- factor(m$type, levels = c("TD", "PD", "FD"))
levels(m$type) <- c("Taxonomic", "Phylogenetic","Functional")

m$predictor = m$variable
remove=c("surveyPeriod1:","surveyPeriod2:")
m$predictor = str_remove_all(m$predictor, paste(remove, collapse = "|"))



m[which(m$predictor=="nInflorescenceCoocc"),c("value")] = 
  m[which(m$predictor=="nInflorescenceCoocc"),c("value")] * 100

m[which(m$predictor=="nInflorescenceCoocc"),c("se_low")] = 
  m[which(m$predictor=="nInflorescenceCoocc"),c("se_low")] * 100

m[which(m$predictor=="nInflorescenceCoocc"),c("se_high")] = 
  m[which(m$predictor=="nInflorescenceCoocc"),c("se_high")] * 100

m = m %>% dplyr::filter(predictor=="nInflorescenceCoocc")

m$predictor = as.factor(m$predictor)
levels(m$predictor) <- c("inflorescences")

percentChangeBeeSRVegetation = m 

# increase of wild bee diversity with 100 more flower
m %>% filter(predictor=="inflorescences",surveyPeriod=="Summer peak") %>% dplyr::summarise(value=mean(value))
```


##### Vegetation plot butterfly
```{r}
fit.TD = modelResultsTDVegetation[[2]]
fit.PD = modelResultsPDVegetation[[2]]
fit.FD = modelResultsFDVegetation[[2]]

fixef.TD <- data.frame(fixef(fit.TD))
fixef.TD$variable <- rownames(fixef.TD)
fixef.TD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","year2022","elevationScaled"))
fixef.TD$type = rep("TD",nrow(fixef.TD))
fixef.TD %<>% dplyr::rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")


fixef.PD <- data.frame(fixef(fit.PD))
fixef.PD$variable <- rownames(fixef.PD)
fixef.PD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","year2022","elevationScaled"))
fixef.PD$type = rep("PD",nrow(fixef.PD))
fixef.PD %<>% dplyr::rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")

fixef.FD <- data.frame(fixef(fit.FD))
fixef.FD$variable <- rownames(fixef.FD)
fixef.FD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","year2022","elevationScaled"))
fixef.FD$type = rep("FD",nrow(fixef.FD))
fixef.FD %<>% dplyr::rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")



m = bind_rows(fixef.TD,fixef.PD,fixef.FD)
rownames(m)= 1:nrow(m)
# 3 diversity types (TD, PD, FD)
# 2 survey periods
m$value   = NA
m$se_low  = NA
m$se_high = NA

m$value= (exp(m$estimate)-1)*100
se=(exp(m$se)-1)*100
m$se_low  = m$value-se
m$se_high = m$value+se

m$surveyPeriod = 1
m[which(grepl("surveyPeriod2",m$variable)),"surveyPeriod"] = 2

# plot
m$surveyPeriod = as.factor(m$surveyPeriod)
m$surveyPeriod = plyr::revalue(m$surveyPeriod, c("1"="Summer onset", "2"="Summer peak"))
m$type = as.factor(m$type)
m$type <- factor(m$type, levels = c("TD", "PD", "FD"))
levels(m$type) <- c("Taxonomic", "Phylogenetic","Functional")

m$predictor = m$variable
remove=c("surveyPeriod1:","surveyPeriod2:")
m$predictor = str_remove_all(m$predictor, paste(remove, collapse = "|"))



m[which(m$predictor=="nInflorescenceCoocc"),c("value")] = 
  m[which(m$predictor=="nInflorescenceCoocc"),c("value")] * 100

m[which(m$predictor=="nInflorescenceCoocc"),c("se_low")] = 
  m[which(m$predictor=="nInflorescenceCoocc"),c("se_low")] * 100

m[which(m$predictor=="nInflorescenceCoocc"),c("se_high")] = 
  m[which(m$predictor=="nInflorescenceCoocc"),c("se_high")] * 100





m = m %>% dplyr::filter(predictor == "nInflorescenceCoocc")

m$predictor = as.factor(m$predictor)
levels(m$predictor) <- c("inflorescences")

percentChangeButterflySRVegetation = m

# increase of butterfly diversity with 100 more flower
m %>% filter(predictor=="inflorescences",surveyPeriod=="Summer peak") %>% dplyr::summarise(value=mean(value))
```


##### Vegetation plot leafhopper
```{r}
fit.TD = modelResultsTDVegetation[[3]]
fit.PD = modelResultsPDVegetation[[3]]
fit.FD = modelResultsFDVegetation[[3]]

fixef.TD <- data.frame(fixef(fit.TD))
fixef.TD$variable <- rownames(fixef.TD)
fixef.TD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","elevationScaled"))
fixef.TD$type = rep("TD",nrow(fixef.TD))
fixef.TD %<>% rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")


fixef.PD <- data.frame(fixef(fit.PD))
fixef.PD$variable <- rownames(fixef.PD)
fixef.PD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","elevationScaled"))
fixef.PD$type = rep("PD",nrow(fixef.PD))
fixef.PD %<>% rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")

fixef.FD <- data.frame(fixef(fit.FD))
fixef.FD$variable <- rownames(fixef.FD)
fixef.FD %<>% dplyr::filter(!variable %in% c("Intercept","surveyPeriod2","elevationScaled"))
fixef.FD$type = rep("FD",nrow(fixef.FD))
fixef.FD %<>% rename(estimate = "Estimate",se="Est.Error",lowCI = "Q2.5", highCI = "Q97.5")



m = bind_rows(fixef.TD,fixef.PD,fixef.FD)
rownames(m)= 1:nrow(m)
# 3 diversity types (TD, PD, FD)
# 2 survey periods
m$value   = NA
m$se_low  = NA
m$se_high = NA

m$value= (exp(m$estimate)-1)*100
se=(exp(m$se)-1)*100
m$se_low  = m$value-se
m$se_high = m$value+se

m$surveyPeriod = 1
m[which(grepl("surveyPeriod2",m$variable)),"surveyPeriod"] = 2

# plot
m$surveyPeriod = as.factor(m$surveyPeriod)
m$surveyPeriod = plyr::revalue(m$surveyPeriod, c("1"="Summer onset", "2"="Summer peak"))
m$type = as.factor(m$type)
m$type <- factor(m$type, levels = c("TD", "PD", "FD"))
levels(m$type) <- c("Taxonomic", "Phylogenetic","Functional")

m$predictor = m$variable
remove=c("surveyPeriod1:","surveyPeriod2:")
m$predictor = str_remove_all(m$predictor, paste(remove, collapse = "|"))



m[which(m$predictor=="biomass"),c("value")] = 
  m[which(m$predictor=="biomass"),c("value")] * 10

m[which(m$predictor=="biomass"),c("se_low")] = 
  m[which(m$predictor=="biomass"),c("se_low")] * 10

m[which(m$predictor=="biomass"),c("se_high")] = 
  m[which(m$predictor=="biomass"),c("se_high")] * 10

m =m %>%dplyr::filter(predictor == "height")
m$predictor = as.factor(m$predictor)
levels(m$predictor) <- c("sward height") # sward height increase by 1 cm

percentChangeLeafhopperSRVegetation = m
```


##### Prepare for Excel export and number of paper
```{r}
percentChangeBeeSRVegetation$group ="Wild bees"
percentChangeButterflySRVegetation$group = "Butterflies"
percentChangeLeafhopperSRVegetation$group = "Leafhoppers"


dat = bind_rows(percentChangeBeeSRVegetation,
                percentChangeButterflySRVegetation,
                percentChangeLeafhopperSRVegetation)
dat$surveyPeriod = as.character(dat$surveyPeriod)
dat[dat$surveyPeriod=="Survey period 1","surveyPeriod"] =1
dat[dat$surveyPeriod=="Survey period 2","surveyPeriod"] =2
resultsExcelPercentChangeSRVegetation = dat
```



## Export to excel
```{r}
wb <- openxlsx::createWorkbook()

openxlsx::addWorksheet(wb, "PercentChangeManagement")
openxlsx::addWorksheet(wb, "PercentChangeVegetation")


openxlsx::writeData(wb, sheet = "PercentChangeManagement",resultsExcelPercentChangeSRManagement)
openxlsx::writeData(wb, sheet = "PercentChangeVegetation",resultsExcelPercentChangeSRVegetation)

openxlsx::saveWorkbook(wb, paste("results/",gsub("-",".",Sys.Date()),
                            "-Q1_modelResults-PercentChange.xlsx",sep=""), overwrite = FALSE)
```

# Combine plot data
Data needed to create Fig. 2
```{r}
dataFig2 = list(
percentChangeBeeSRManagement = percentChangeBeeSRManagement,
percentChangeButterflySRManagement = percentChangeButterflySRManagement,
percentChangeLeafhopperSRManagement = percentChangeLeafhopperSRManagement,
percentChangeBeeSRVegetation = percentChangeBeeSRVegetation,
percentChangeButterflySRVegetation = percentChangeButterflySRVegetation,
percentChangeLeafhopperSRVegetation = percentChangeLeafhopperSRVegetation)

save(dataFig2,file="data/dataFigures_2_GLMM_percent_change.RData")
```





# Save model results
```{r}
save(resultsExcelPercentChangeSRManagement,
     resultsExcelPercentChangeSRVegetation,percentChangeVeg,
     file="results/modelResults-PercentChange.RData")
```
