==> r-code/03_Q2_1_statistical_analyses_SEM.Rmd <==
---
title: "SEMs to evaluate the influence of grazing and environment on insect diversity"
author: "Bernd Panassiti"
date: 'created: 07.06.2023, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output:
  pdf_document: default
  html_document: default
editor_options: 
  chunk_output_type: console
---



```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```




# Introduction
Useful link:
https://rpubs.com/jebyrnes/brms_bayes_sem



# Loading of packages and data
```{r loading_data, include=FALSE}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load(tidyverse,magrittr,caret,ggplot2,DHARMa,broom.mixed,
       posterior,ggdist,tidybayes,pryr,bayesplot,ggpubr,grid,
       gridExtra,ggdag,dagitty,
       emmeans,broom, broom.mixed,
       posterior,brms,bayesplot,ggpubr)

source("r-code/00_functions.R")

load("data/workingdata.RData")
load("results/iNEXT3d_bee.RData")
load("results/iNEXT3d_butterfly.RData")
load("results/iNEXT3d_planthopper.RData")
```





# Data preparation

```{r}
dfbee = beeDiversity%>% tibble::add_column(group = rep("bee",nrow(beeDiversity)))
dfbutterfly = butterflyDiversity%>% tibble::add_column(group = rep("butterfly",nrow(butterflyDiversity)))
dfplanthopper = planthopperDiversity%>% tibble::add_column(group = rep("planthopper",nrow(planthopperDiversity)))

dfInsects = bind_rows(dfbee,dfbutterfly,dfplanthopper) 
dfInsects %<>%
  dplyr::rename(grazing_time = grazingTime) %>%
  dplyr::mutate(response = paste(group,diversityType,"q",qHill,sep="")) %>%
  dplyr::select(plot_id,site_id, survey_period,year,grazing_time,response,estimate)


dfIns = reshape::cast(dfInsects,plot_id+site_id+survey_period+year+grazing_time ~response)
```


```{r}
dfVeg = veg %>% rename(nPlantSpecies = nSpecies)

dfInPl = merge(dfIns,dfVeg,
               by=c("site_id","year","survey_period","grazing_time"),all.x = T)
data = merge(dfInPl,sitesTime,
              by.y=c("site_id","grazing_time"),all.x = T)

data$grazing_time <- factor(data$grazing_time, levels = c("control", "early", "late"))
data$survey_period = as.factor(data$survey_period)
data$year = as.factor(data$year)

data$nInflorescenceCooccScaled = as.numeric(scale(data$nInflorescenceCooc))
data$nPlantSpeciesScaled = as.numeric(scale(data$nPlantSpecies))
data$coverHerbsScaled = as.numeric(scale(data$coverHerbs))
data$heightScaled = as.numeric(scale(data$height))
data$biomassScaled = as.numeric(scale(data$biomass))
data$sugarScaled = as.numeric(scale(data$sugar))
data$elevationScaled = as.numeric(scale(data$elevation))

# rename columns to avoid underscore
data %<>%  dplyr::rename(siteID = site_id,
                surveyPeriod = survey_period,
                grazingTime = grazing_time)
model_data = data
```


# Piecewise SEM models

## Settings
```{r}
## Specify model parameters
family=gaussian()
myIter = 30000 # number of MCMC samples
myThin = 10 # thinning in case of strong autocorrelation of Markov chains
myChains = 4 # number of Markov chains
mySeed = 2022 # sets the internal random number generator seed to allow reproducibility
# control: tuning parameter in the NUTS sampler for Hamiltonian Monte Carlo, can be raised in case of divergent transitions, btw 0 and 1, default: 0.8
myPriors = c(prior(normal(0,10),class="Intercept"),
             prior(normal(0,10),class="b"))
controlList = list(adapt_delta = 0.9999, max_treedepth = 15)

```


## Fitting the models
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2023)

make_model <- function(y) brms::brm(
  y[['iFormula']]+y[['pFormula']]+set_rescor(FALSE), 
  data=model_data,
              family=family,iter = myIter, thin = myThin, chains = myChains,
              seed=mySeed, control=controlList,
              save_pars = save_pars(all = TRUE))


# TD
model_specs <- list(
# bee
bee.TD.q0 = list(
iFormula = brmsformula(beeTDq0 ~ surveyPeriod+ 
                         grazingTime:surveyPeriod+nInflorescenceCoocc+
                         year+elevationScaled+ (1|siteID)),
pFormula = brmsformula(nInflorescenceCoocc ~ surveyPeriod + 
                         grazingTime:surveyPeriod +
                         year+elevationScaled+ (1|siteID))),

#butterfly
butterfly.TD.q0 = list(iFormula = brmsformula(butterflyTDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+nInflorescenceCoocc+
                                            year+elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(nInflorescenceCoocc ~ surveyPeriod + grazingTime:surveyPeriod +
                                           year+elevationScaled+ (1 |siteID))),

#planthopper
planthopper1.TD.q0 = list(iFormula = brmsformula(planthopperTDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+nPlantSpecies+
                                            elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(nPlantSpecies ~ surveyPeriod + grazingTime:surveyPeriod +
                                           elevationScaled+ (1 |siteID))),

planthopper2.TD.q0 = list(iFormula = brmsformula(planthopperTDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+height+
                                            elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(height ~ surveyPeriod + grazingTime:surveyPeriod +
                                           elevationScaled+ (1 |siteID))),
planthopper3.TD.q0 = list(iFormula = brmsformula(planthopperTDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+biomass+
                                            elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(biomass ~ surveyPeriod + grazingTime:surveyPeriod +
                                           elevationScaled+ (1 |siteID)))

)

modelResultsTDVegetationManagement <- model_specs  %>% Map(f = make_model) 

# PD
model_specs <- list(
# bee
bee.PD.q0 = list(iFormula = brmsformula(beePDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+nInflorescenceCoocc+
                                            year+elevationScaled+ (1 |siteID)),
                   pFormula = brmsformula(nInflorescenceCoocc ~ surveyPeriod + grazingTime:surveyPeriod +
                                           year+elevationScaled+ (1 |siteID))),
#butterfly
butterfly.PD.q0 = list(iFormula = brmsformula(butterflyPDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+nInflorescenceCoocc+
                                            year+elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(nInflorescenceCoocc ~ surveyPeriod + grazingTime:surveyPeriod +
                                           year+elevationScaled+ (1 |siteID))),
#planthopper
planthopper1.PD.q0 = list(iFormula = brmsformula(planthopperPDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+nPlantSpecies+
                                            elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(nPlantSpecies ~ surveyPeriod + grazingTime:surveyPeriod +
                                            elevationScaled+ (1 |siteID))),

planthopper2.PD.q0 = list(iFormula = brmsformula(planthopperPDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+height+
                                            elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(height ~ surveyPeriod + grazingTime:surveyPeriod +
                                            elevationScaled+ (1 |siteID))),

planthopper3.PD.q0 = list(iFormula = brmsformula(planthopperPDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+biomass+
                                            elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(biomass ~ surveyPeriod + grazingTime:surveyPeriod +
                                            elevationScaled+ (1 |siteID))))

modelResultsPDVegetationManagement <- model_specs  %>% Map(f = make_model) 


# FD
model_specs <- list(
# bee
bee.FD.q0 = list(iFormula = brmsformula(beeFDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+nInflorescenceCoocc+
                                            year+elevationScaled + (1 |siteID)),
                   pFormula = brmsformula(nInflorescenceCoocc ~ surveyPeriod + grazingTime:surveyPeriod +
                                           year+elevationScaled + (1 |siteID))),
#butterfly
butterfly.FD.q0 = list(iFormula = brmsformula(butterflyFDq0 ~ surveyPeriod+ grazingTime:surveyPeriod+nInflorescenceCoocc+
                                            year+elevationScaled + (1 |siteID)),
                       pFormula = brmsformula(nInflorescenceCoocc ~ surveyPeriod + grazingTime:surveyPeriod +
                                           year+elevationScaled + (1 |siteID))),

#planthopper
planthopper1.FD.q0 = list(iFormula = brmsformula(planthopperFDq0 ~ surveyPeriod+
                                            grazingTime:surveyPeriod+nPlantSpecies+
                                            elevationScaled+(1|siteID)),
                       pFormula = brmsformula(nPlantSpecies ~ surveyPeriod +
                                            grazingTime:surveyPeriod + 
                                              elevationScaled+(1|siteID))),

planthopper2.FD.q0 = list(iFormula = brmsformula(planthopperFDq0 ~ surveyPeriod+
                                          grazingTime:surveyPeriod+height+
                                            elevationScaled+(1|siteID)),
                       pFormula = brmsformula(height ~ surveyPeriod +
                                        grazingTime:surveyPeriod +
                                        elevationScaled+(1|siteID))),

planthopper3.FD.q0 = list(iFormula = brmsformula(planthopperFDq0 ~surveyPeriod+
                                                   grazingTime:surveyPeriod+biomass+
                                                   elevationScaled+ (1 |siteID)),
                       pFormula = brmsformula(biomass ~ surveyPeriod +
                                            grazingTime:surveyPeriod +
                                            elevationScaled+(1|siteID))))

modelResultsFDVegetationManagement <- model_specs  %>% Map(f = make_model) 
```


## Plots

++++### Esimates
https://cran.r-project.org/web/packages/bayesplot/vignettes/plotting-mcmc-draws.html

https://bookdown.org/content/c23e1ece-b5d6-4cab-8d8e-978c2e5b7a53/multivariate-linear-models.html

https://nicholasrjenkins.science/post/bda_in_r/bayes_with_brms/

https://discourse.mc-stan.org/t/manually-changing-y-axis-tick-labels-in-brms-mcmc-plot/19727

```{r}
#TD
n = names(modelResultsTDVegetationManagement)

plot_list = lapply(setNames(n, n), function(nameindex) {mcmc_plot(modelResultsTDVegetationManagement[[nameindex]],
                variable = c( "^b_", "^sd_"), regex = T,prob_outer = 0.95) +
                labs(title = gsub("[.]"," ",nameindex), tag = "")})



for (i in 1:3){
Labels = as.character(plot_list[[i]]$data$parameter)
Labels = gsub("b_","",Labels)
Labels = gsub("beeTDq0","i",Labels)
Labels = gsub("butterflyTDq0","i",Labels)
Labels = gsub("planthopperTDq0","i",Labels)
Labels = gsub("nInflorescenceCoocc","p",Labels)
Labels = gsub("nPlantSpecies","p",Labels)
Labels[which(Labels =="sd_siteID__i_Intercept")] ="i_sd siteID (Intercept)"
Labels[which(Labels =="sd_siteID__p_Intercept")] ="p_sd siteID (Intercept)"
Labels= rev(Labels)
plot_list[[i]] = plot_list[[i]] + scale_y_discrete(labels= Labels, limits = rev)
}




for (i in c(2,3)){
plot_list[[i]] = plot_list[[i]]+
           theme(axis.ticks.y = element_blank(),
                 axis.text.y=element_blank())}

plotSEM_TD = ggpubr::ggarrange(plotlist=plot_list,widths = c(1, 0.5,0.5),ncol=3);plotSEM_TD 



# PD
n = names(modelResultsPDVegetationManagement)

plot_list = lapply(setNames(n, n), function(nameindex) {mcmc_plot(modelResultsPDVegetationManagement[[nameindex]],
                variable = c( "^b_", "^sd_"), regex = T,
                prob_outer = 0.95) +
                labs(title = gsub("[.]"," ",nameindex), tag = "")})



for (i in 1:3){
Labels = as.character(plot_list[[i]]$data$parameter)
Labels = gsub("b_","",Labels)
Labels = gsub("beePDq0","i",Labels)
Labels = gsub("butterflyPDq0","i",Labels)
Labels = gsub("planthopperPDq0","i",Labels)
Labels = gsub("nInflorescenceCoocc","p",Labels)
Labels = gsub("nPlantSpecies","p",Labels)
Labels[which(Labels =="sd_siteID__i_Intercept")] ="i_sd siteID (Intercept)"
Labels[which(Labels =="sd_siteID__p_Intercept")] ="p_sd siteID (Intercept)"
Labels= rev(Labels)
plot_list[[i]] = plot_list[[i]] + scale_y_discrete(labels= Labels, limits = rev)
}

for (i in c(2,3)){
plot_list[[i]] = plot_list[[i]]+
           theme(axis.ticks.y = element_blank(),
                 axis.text.y=element_blank())}

plotSEM_PD = ggpubr::ggarrange(plotlist=plot_list,widths = c(1, 0.5,0.5),ncol=3);plotSEM_PD


# FD
n = names(modelResultsFDVegetationManagement)

plot_list = lapply(setNames(n, n), function(nameindex) {mcmc_plot(modelResultsFDVegetationManagement[[nameindex]],
                variable = c( "^b_", "^sd_"), regex = T,prob_outer = 0.95) +
                labs(title = gsub("[.]"," ",nameindex), tag = "")})



for (i in 1:3){
Labels = as.character(plot_list[[i]]$data$parameter)
Labels = gsub("b_","",Labels)
Labels = gsub("beeFDq0","i",Labels)
Labels = gsub("butterflyFDq0","i",Labels)
Labels = gsub("planthopperFDq0","i",Labels)
Labels = gsub("nInflorescenceCoocc","p",Labels)
Labels = gsub("nPlantSpecies","p",Labels)
Labels[which(Labels =="sd_siteID__i_Intercept")] ="i_sd siteID (Intercept)"
Labels[which(Labels =="sd_siteID__p_Intercept")] ="p_sd siteID (Intercept)"
Labels= rev(Labels)
plot_list[[i]] = plot_list[[i]] + scale_y_discrete(labels= Labels, limits = rev)
}

for (i in c(2,3)){
plot_list[[i]] = plot_list[[i]]+
           theme(axis.ticks.y = element_blank(),
                 axis.text.y=element_blank())}

plotSEM_FD = ggpubr::ggarrange(plotlist=plot_list,widths = c(1, 0.5,0.5),ncol=3);plotSEM_FD
```


# Export in excel table

## Estimates
```{r}
modelEstimatesTDVegetationManagement = modelResultsTDVegetationManagement %>% lapply(function(x) broom::tidy(x))
datTD = list_rbind(modelEstimatesTDVegetationManagement, names_to = "id")

modelEstimatesPDVegetationManagement = modelResultsPDVegetationManagement %>% lapply(function(x) broom::tidy(x))
datPD = list_rbind(modelEstimatesPDVegetationManagement, names_to = "id")

modelEstimatesFDVegetationManagement = modelResultsFDVegetationManagement %>% lapply(function(x) broom::tidy(x))
datFD = list_rbind(modelEstimatesFDVegetationManagement, names_to = "id")

df = bind_rows(datTD,datPD,datFD)


# Intercept surveyPeriod2 year2022 elevation sp1xEarly sp2xEarly sp1xLate sp2xLate


df  %<>% tibble::add_column(effectNoteworthy=NA)


for(i in 1:nrow(df)){
  if(isTRUE(data.table::between(0,df$conf.low[i],df$conf.high[i]))){
   df$effectNoteworthy[i] = ""}else{
   df$effectNoteworthy[i] = "*"}
}


df %<>% mutate(estimateNew = paste(my_signif(estimate)," (",my_signif(std.error),")",effectNoteworthy,sep=""),
                   conf = paste("(",my_signif(conf.low),", ",my_signif(conf.high),")",sep="")
                   )

# rename responses
df[which(grepl("bee",df$response)),"response"] = "i"
df[which(grepl("butterfly",df$response)),"response"] = "i"
df[which(grepl("planthopper",df$response)),"response"] = "i"
df[which(grepl("nInflorescenceCoocc",df$response)),"response"] = "p"
df[which(grepl("nPlantSpecies",df$response)),"response"] = "p"


df[which(grepl("siteID",df$response) & grepl("bee",df$group)),"response"] = "i"
df[which(grepl("siteID",df$response) & grepl("butterfly",df$group)),"response"] = "i"
df[which(grepl("siteID",df$response) & grepl("planthopper",df$group)),"response"] = "i"
df[which(grepl("siteID",df$response) & grepl("nInflorescenceCoocc",df$group)),"response"] = "p"
df[which(grepl("siteID",df$response) & grepl("nPlantSpecies",df$group)),"response"] = "p"

df = df %>% dplyr::filter(effect=="fixed")

df %<>%dplyr::select(-effect,-component,-group, -estimate,-std.error,-conf.low,-conf.high,-effectNoteworthy)


df1 = df %>% pivot_wider(id_cols = c("id","response"),names_from = term,values_from = c("estimateNew","conf"))

# renaming for columns
names(df1) =  gsub("estimateNew_","",names(df1))

df2 = df1%>%dplyr::rename(
  Intercept = '(Intercept)',
  inflor = 'nInflorescenceCoocc',
  SRplant = 'nPlantSpecies',
  sp1xEarly='surveyPeriod1:grazingTimeearly',
  sp2xEarly='surveyPeriod2:grazingTimeearly',
  sp1xLate='surveyPeriod1:grazingTimelate',
  sp2xLate='surveyPeriod2:grazingTimelate',
  #sd_Intercept = 'sd__(Intercept)',
  #sd_Obs = 'sd__Observation',
  conf_Intercept = 'conf_(Intercept)',
  conf_sp1xEarly = 'conf_surveyPeriod1:grazingTimeearly',
  conf_sp2xEarly = 'conf_surveyPeriod2:grazingTimeearly',
  conf_sp1xLate = 'conf_surveyPeriod1:grazingTimelate',
  conf_sp2xLate = 'conf_surveyPeriod2:grazingTimelate',
  conf_inflor =  'conf_nInflorescenceCoocc',
  conf_SRplant = 'conf_nPlantSpecies',
  #conf_sd_Intercept = 'conf_sd__(Intercept)',              
  #conf_sd_Obs = 'conf_sd__Observation'              
)

# final noteworthy table
df2  %<>% tibble::add_column(effectNoteworthy=0)


for(i in 1:nrow(df2)){
  if(any(sapply(df2[i,3:12], function(x) grepl("\\*",x)))){
   df2$effectNoteworthy[i] = 1}
}
```


## Performance
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2023)

bayesRTDVegetationManagement = modelResultsTDVegetationManagement %>% lapply(function(x) brms::bayes_R2(x))
looTDVegetationManagement = modelResultsTDVegetationManagement %>% lapply(function(x) brms::loo(x, save_psis = TRUE, k_threshold = 0.7,reloo=T))

bayesRPDVegetationManagement = modelResultsPDVegetationManagement %>% lapply(function(x) brms::bayes_R2(x))
looPDVegetationManagement = modelResultsPDVegetationManagement %>% lapply(function(x) brms::loo(x, save_psis = TRUE, k_threshold = 0.7,reloo=T))

bayesRFDVegetationManagement = modelResultsFDVegetationManagement %>% lapply(function(x) brms::bayes_R2(x))
looFDVegetationManagement = modelResultsFDVegetationManagement %>% lapply(function(x) brms::loo(x, save_psis = TRUE, k_threshold = 0.7,reloo=T))




# Response residuals not available to calculate mean square error. (R)MSE is probably not reliable.

dfModelbayesRTDVegetationManagement = as_tibble(do.call(rbind, bayesRTDVegetationManagement))
dfModelbayesRTDVegetationManagement$response = rep(c("i","p"),5)
dfModelbayesRTDVegetationManagement$id = rep(names(bayesRTDVegetationManagement),each=2)

dfModelbayesRPDVegetationManagement = as_tibble(do.call(rbind, bayesRPDVegetationManagement))
dfModelbayesRPDVegetationManagement$response = rep(c("i","p"),5)
dfModelbayesRPDVegetationManagement$id = rep(names(bayesRPDVegetationManagement),each=2)

dfModelbayesRFDVegetationManagement = as_tibble(do.call(rbind, bayesRFDVegetationManagement))
dfModelbayesRFDVegetationManagement$response = rep(c("i","p"),5)
dfModelbayesRFDVegetationManagement$id = rep(names(bayesRFDVegetationManagement),each=2)


dfModelbayesRTDVegetationManagement = as_tibble(do.call(rbind, bayesRTDVegetationManagement))
dfModelbayesRTDVegetationManagement$response = rep(c("i","p"),5)
dfModelbayesRTDVegetationManagement$id = rep(names(bayesRTDVegetationManagement),each=2)

dfModelbayesRPDVegetationManagement = as_tibble(do.call(rbind, bayesRPDVegetationManagement))
dfModelbayesRPDVegetationManagement$response = rep(c("i","p"),5)
dfModelbayesRPDVegetationManagement$id = rep(names(bayesRPDVegetationManagement),each=2)

dfModelbayesRFDVegetationManagement = as_tibble(do.call(rbind, bayesRFDVegetationManagement))
dfModelbayesRFDVegetationManagement$response = rep(c("i","p"),5)
dfModelbayesRFDVegetationManagement$id = rep(names(bayesRFDVegetationManagement),each=2)


bayesrResult = bind_rows(dfModelbayesRTDVegetationManagement,
                                         dfModelbayesRPDVegetationManagement ,
                                         dfModelbayesRFDVegetationManagement)



looTD = lapply(looTDVegetationManagement, function(x) x$estimates[1,1])
looTDdf = data.frame(unlist(looTD));colnames(looTDdf) = "elpd_loo"
looPD = lapply(looPDVegetationManagement, function(x) x$estimates[1,1])
looPDdf = data.frame(unlist(looPD));colnames(looPDdf) = "elpd_loo"
looFD = lapply(looFDVegetationManagement, function(x) x$estimates[1,1])
looFDdf = data.frame(unlist(looFD));colnames(looFDdf) = "elpd_loo"

looResult = rbind(looTDdf,looPDdf,looFDdf)
looResult$id = rownames(looResult)
```

## Prepare for Excel-Export

```{r}
dfBayes= bayesrResult %>% dplyr::select(id,response,Estimate) %>% 
  dplyr::rename(bayesR = Estimate) 


dfTemp = merge(df2,dfBayes,by=c("id","response"))
df3 =  merge(dfTemp,looResult,by="id")

df3$group = as.character(sapply(df3$id, function(x) str_split(x,"\\.")[[1]][1]))
df3$divType = as.character(sapply(df3$id, function(x) 
                str_split(x,"\\.")[[1]][2]))
df3$hillNumber = as.numeric(sapply(df3$id, function(x)
                  gsub("q","",str_split(x,"\\.")[[1]][3])))

df3 %<>% dplyr::select(-id)
resultsExcel = arrange.vars(df3, c("group"=1, "divType"=2, "hillNumber"=3))

resultsExcel[,c("bayesR","elpd_loo")] = apply(resultsExcel[,c("bayesR","elpd_loo")],2,my_signif)


resultsExcelSEM = as_tibble(resultsExcel)
```


# Table for manuscript
Table contains bayes R and elpd loo values
```{r}
df = bayesrResult

# create new column modelType
df$modelType <- sub("\\..*", "", df$id)

# delete entries for planthopper1 and planthopper3 
# -> only planthopper2 needed, model uses sward height
df %<>% dplyr::filter(!modelType %in% c("planthopper1","planthopper3"))


# create column diversityType with values TD, PD, and FD
df$diversityType = sub("^[^.]*\\.([^.]*)\\..*$", "\\1", df$id)

# rename category planthopper2 in leafhopper
df %<>% dplyr::mutate(modelType = recode(modelType, "planthopper2" = "Leafhopper"))
df %<>% dplyr::mutate(modelType = recode(modelType, "butterfly" = "Butterfly"))
df %<>% dplyr::mutate(modelType = recode(modelType, "bee" = "Wild bee"))

# rename category planthopper2 in leafhopper
df %<>% mutate(response = ifelse(response == "p" & 
                                   modelType %in% c("Wild bee", "Butterfly"), "Number of inflorescences", response))

df %<>% mutate(response = ifelse(response == "p" & 
                                   modelType %in% c("Leafhopper"), "Sward height", response))

df %<>% mutate(response = ifelse(response == "i" & 
                                   modelType %in% c("Wild bee"), "Wild bee SR", response))

df %<>% mutate(response = ifelse(response == "i" & 
                                   modelType %in% c("Butterfly"), "Butterfly SR", response))

df %<>% mutate(response = ifelse(response == "i" & 
                                   modelType %in% c("Leafhopper"), "Leafhopper SR", response))


df %<>% dplyr::select(-id)
df %<>% select(modelType, diversityType, response, everything())


df$Estimate  = round(df$Estimate,2)
df$Q2.5      = round(df$Q2.5,2)
df$Q97.5     = round(df$Q97.5,2)
df$Est.Error = my_signif(df$Est.Error,3)


resultsExcelSEMbayesR = df
```



## Export to excel
```{r}
openxlsx::write.xlsx(resultsExcelSEM, paste("results/",gsub("-",".",Sys.Date()),
                            "-Q2_modelResultsSEM.xlsx",sep=""))

openxlsx::write.xlsx(resultsExcelSEMbayesR, paste("results/",gsub("-",".",Sys.Date()),
                            "-Q2_modelResultsSEM_bayesR.xlsx",sep=""))
```



# Save model results
```{r}
save(model_data, 
     modelResultsTDVegetationManagement,
     modelResultsPDVegetationManagement,
     modelResultsFDVegetationManagement,
     resultsExcelSEM ,bayesrResult,looResult,
     resultsExcelSEMbayesR,
     file="results/modelResultsSEM.RData")
```
