==> r-code/03_Q3_2_iNextlink.Rmd <==
---
title: "iNEXT link - Plots for plant-insect interactions"
author: ""
date: 'created: 20.01.2022, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output:
  pdf_document: default
  html_document: default
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')                        # pdf device
```



```{r}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load("tidyverse","vegan","reshape2","magrittr","caret","ggplot2",
       "ape","reshape2","phytools","bipartite","patchwork","iNEXT.link","StatMatch","egg")

source("r-code/00_functions.R")
load("data/workingdata.RData")

load("results/Q2_bee_iNext.link.RData")
load("results/Q2_butterfly_iNext.link.RData")
load("results/Q2_planthopper_iNext.link.RData")
```


```{r}
ggpubr::ggarrange(
  bee.cov.TD +ylab("Taxonomic\nnetwork diversity") + 
    ggtitle("Wild bees")+
    theme(plot.title = element_text(colour = "black")),
  butterfly.cov.TD + ylab("") +
    ggtitle("Butterflies") +
    theme(plot.title = element_text(colour = "black")),
  planthopper.cov.TD + ylab("") +
    expand_limits(y = 0) +
    ggtitle("Leafhoppers") +
    theme(plot.title = element_text(colour = "black")),
  
  bee.cov.PD +ylab("Phylogenetic\nnetwork diversity") +
    ggtitle(""),
  butterfly.cov.PD + ggtitle("") + ylab(""),
  planthopper.cov.PD + ggtitle("") + ylab("") +
    coord_cartesian(xlim = c(0.9,1)),
  
  bee.cov.FD +ylab("Functional\nnetwork diversity") +
    ggtitle(""),
  butterfly.cov.FD + ggtitle("") + ylab(""),
  planthopper.cov.FD + ggtitle("") + ylab(""),
  
  ncol=3,nrow=3,
  common.legend = T,
  legend = "bottom",
  labels = "auto")  
```


# Coverage

```{r}
## Get observed
# bee
df1 = bee.iNEXT.TD$TDiNextEst$coverage_based %>% dplyr::filter(Method=="Observed") 
df2 = bee.est.TD %>% dplyr::select(-s.e.)
beeEstTD= bind_rows(df1,df2)

df1 = bee.iNEXT.PD$PDiNextEst$coverage_based %>% dplyr::filter(Method=="Observed") %>% 
  dplyr::relocate(Order.q, .before = SC)
df2 = bee.est.PD %>% dplyr::select(-s.e.)
beeEstPD= bind_rows(df1,df2)


df1 = bee.iNEXT.FD$FDiNextEst$coverage_based %>% dplyr::filter(Method=="Observed") 
df2 = bee.est.FD %>% dplyr::select(-s.e.)
beeEstFD= bind_rows(df1,df2)


# butterfly
df1 = butterfly.iNEXT.TD$TDiNextEst$coverage_based %>% dplyr::filter(Method=="Observed") 
df2 = butterfly.est.TD %>% dplyr::select(-s.e.)
butterflyEstTD= bind_rows(df1,df2)

df1 = butterfly.iNEXT.PD$PDiNextEst$coverage_based %>% dplyr::filter(Method=="Observed") %>% 
  dplyr::relocate(Order.q, .before = SC)
df2 = butterfly.est.PD %>% dplyr::select(-s.e.)
butterflyEstPD= bind_rows(df1,df2)


df1 = butterfly.iNEXT.FD$FDiNextEst$coverage_based %>% dplyr::filter(Method=="Observed") 
df2 = butterfly.est.FD %>% dplyr::select(-s.e.)
butterflyEstFD= bind_rows(df1,df2)

planthopperEstTD=planthopper.est.TD
planthopperEstPD=planthopper.est.PD
planthopperEstFD=planthopper.est.FD
```

```{r}
netResults = data.frame(matrix(ncol=8,nrow=9))
colnames(netResults) = c("group","type","c_lower_e","c_lower_l","c_higher_e","c_higher_l","e_lower_l","l_lower_e")
netResults$group = c(rep("Wildbees",3),rep("Butterflies",3),rep("Leafhoppers",3))
netResults$type = rep(c("TD","PD","FD"),3)

beeEstTD$group = "Wildbees"
beeEstPD$group = "Wildbees"
beeEstFD$group = "Wildbees"

butterflyEstTD$group = "Butterflies"
butterflyEstPD$group = "Butterflies"
butterflyEstFD$group = "Butterflies"

planthopperEstTD$group = "Leafhoppers"
planthopperEstPD$group = "Leafhoppers"
planthopperEstFD$group = "Leafhoppers"

NetTD = bind_rows(beeEstTD,butterflyEstTD,planthopperEstTD)
colnames(NetTD)[which(colnames(NetTD) =="qTD.LCL")] = "low"
colnames(NetTD)[which(colnames(NetTD) =="qTD.UCL")] = "high"
colnames(NetTD)[which(colnames(NetTD) =="qTD")] = "estimate"


NetPD = bind_rows(beeEstPD,butterflyEstPD,planthopperEstPD)
colnames(NetPD)[which(colnames(NetPD) =="qPD.LCL")] = "low"
colnames(NetPD)[which(colnames(NetPD) =="qPD.UCL")] = "high"
colnames(NetPD)[which(colnames(NetPD) =="qPD")] = "estimate"

NetFD = bind_rows(beeEstFD,butterflyEstFD,planthopperEstFD)
colnames(NetFD)[which(colnames(NetFD) =="qFD.LCL")] = "low"
colnames(NetFD)[which(colnames(NetFD) =="qFD.UCL")] = "high"
colnames(NetFD)[which(colnames(NetFD) =="qFD")] = "estimate"

NetTD$type ="TD"
NetPD$type ="PD"
NetFD$type ="FD"

NetTD %<>% dplyr::select(group,type,Assemblage,SC,Method,estimate,low,high) %>%
  dplyr::rename(treatment = Assemblage)

NetPD %<>% dplyr::select(group,type,Assemblage,SC,Method,estimate,low,high) %>%
  dplyr::rename(treatment = Assemblage)

NetFD %<>% dplyr::select(group,type,Assemblage,SC,Method,estimate,low,high) %>%
  dplyr::rename(treatment = Assemblage)

netEstimates = bind_rows(NetTD,NetPD,NetFD)

netEstimates$group <- factor(netEstimates$group, levels = c("Wildbees", "Butterflies", "Leafhoppers"))
netEstimates$type <- factor(netEstimates$type, levels = c("TD", "PD", "FD"))


z = 0
for (i in c("Wildbees","Butterflies","Leafhoppers")){
dat = netEstimates[which(netEstimates$group==i),]

for(j in c("TD","PD","FD")){
  df = dat[which(dat$type==j),]

z = z+1

if(i == "Wildbees"){
df %<>% dplyr::filter(SC >0 & SC<=0.8)
}

if(i == "Butterflies"){
df %<>% dplyr::filter(SC >0 & SC<=0.7)
}

if(i == "Planthoppers"){
df %<>% dplyr::filter(SC >0 & SC<=0.7)
}



control = df %>% dplyr::filter(treatment =="control")
early = df %>% dplyr::filter(treatment =="early")
late = df %>% dplyr::filter(treatment =="late")

# control smaller than treatments
netResults$c_lower_e[z] = any(control$high < early$low)
netResults$c_lower_l[z] = any(control$high < late$low)

netResults$c_higher_e[z] = any(control$low > early$high)
netResults$c_higher_l[z] = any(control$low > late$high)

netResults$e_lower_l[z] = any(early$high < late$low)
netResults$l_lower_e[z] = any(late$high  < early$low)
print(z)
}}
```



# Plot preparation
dfPlot = netEstimates %>% dplyr::filter(group=="Wildbees", type=="PD")
dfLines = dfPlot %>%dplyr::filter(Method!="Observed")
dfPoints = dfPlot %>%dplyr::filter(Method=="Observed")

ggplot(dfPlot, aes(x=SC,ymin=low,ymax=high,fill=treatment))+
  geom_ribbon(alpha=.5)+
  geom_line(data=dfLines,aes(x=SC,y=estimate,color=treatment,linetype=Method),size=1)+
  scale_linetype_manual(values=c( "dotted","solid")) +
  geom_point(data = dfPoints,aes(x=SC,y=estimate), size=2, color="black")


```{r}
# vertical lines
vLineDf = data.frame(matrix(ncol=1,nrow=9))
vLineDf$group = c(rep("(A) Wildbees",3),rep("(B) Butterflies",3),rep("(C) Leafhoppers",3))
vLineDf$type = rep(c("TDnet","PDnet","FDnet"),3)
colnames(vLineDf)[1] = "x_value"

#vLineDf$x_value[1:3] = 0.76
#vLineDf$x_value[4:6] = 0.49
#vLineDf$x_value[7:9] = 1

vLineDf$x_value = 1

vLineDf$group <- factor(vLineDf$group, levels = c("Wild bees", "Butterflies", "Planthoppers"))
vLineDf$type <- factor(vLineDf$type, levels = c("TDnet", "PDnet", "FDnet"))


# points and lines dataframes
dfPlot = netEstimates

dfPlot$group = as.character(dfPlot$group)
dfPlot$type = as.character(dfPlot$type)

dfPlot$group[which(dfPlot$group == "Wildbees")] = "Wild bees"
dfPlot$group[which(dfPlot$group == "Butterflies")] = "Butterflies"
dfPlot$group[which(dfPlot$group == "Planthoppers")] = "Planthoppers"

dfPlot$type[which(dfPlot$type == "TD")] = "TDnet"
dfPlot$type[which(dfPlot$type == "PD")] = "PDnet"
dfPlot$type[which(dfPlot$type == "FD")] = "FDnet"

dfPlot$group <- factor(dfPlot$group, levels = c("Wild bees", "Butterflies", "Leafhoppers"))
dfPlot$type <- factor(dfPlot$type, levels = c("TDnet", "PDnet", "FDnet"))

dataFig4 = dfPlot
```


```{r}
save(
planthopper.est.TD,
planthopper.est.PD,
planthopper.est.FD,
netEstimates,
dataFig4,
file="results/iNext.link-Plots.RData"
)
```
